Fix int overflow in blob_url_request_job when we read files > 2GB
BUG=131295
TEST=manual; successfully copied 3GB file in Chromeos filebrowser


Review URL: https://chromiumcodereview.appspot.com/10536062

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@141181 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: e7d5d9c69d1634ca71fd9ff6f1e27b6fa0b47afc

+++ webkit/blob/blob_url_request_job.cc
@@ -414,8 +414,10 @@ int BlobURLRequestJob::ComputeBytesToRead() const {
       item_length_list_[current_item_index_] - current_item_offset_;
   int64 remaining_bytes = std::min(current_item_remaining_bytes,
                                    remaining_bytes_);
-  return std::min(read_buf_->BytesRemaining(),
-                  static_cast<int>(remaining_bytes));
+
+  return static_cast<int>(std::min(
+             static_cast<int64>(read_buf_->BytesRemaining()),
+             remaining_bytes));
 }
 
 bool BlobURLRequestJob::ReadLoop(int* bytes_read) {
